Add parsing tests to compiler benchmarks
Adds support for testing parsing to the compiler benchmark suite and
refactors the whole suite to make it smaller and simpler.
