+++ src/Tools/CompilerBenchmarks/Program.cs
@@ -22,7 +22,7 @@ namespace CompilerBenchmarks
                 Add(DefaultConfig.Instance.GetExporters().ToArray());
                 Add(DefaultConfig.Instance.GetColumnProviders().ToArray());
                 Add(MemoryDiagnoser.Default);
-                Add(Job.Core);
+                Add(Job.Core.WithGcServer(true));
             }
         }
 
@@ -39,7 +39,7 @@ namespace CompilerBenchmarks
             // to communicate information is pass by environment variable
             Environment.SetEnvironmentVariable(Helpers.TestProjectEnvVarName, projectPath);
 
-            _ = BenchmarkRunner.Run<EmitBenchmark>(config);
+            _ = BenchmarkRunner.Run<StageBenchmarks>(config);
         }
     }
 }
